refactor(clean-arch): move services report/cache/subprocess I/O behind data adapters
Moves the analysis-layer dimension-report cluster and the V2 result-cache
filesystem backend to data/fs/dimension_report and data/cache_store (with
re-export shims at their old analysis/ paths), and gives services new data
adapters for project-dir removal, scan.json reads, cached score-row bulk
reads, CLI /models subprocess discovery, and the Anthropic models HTTP
call. Adds StoreUnreadableError (data/ports/errors.py) so the scoring
facade stops importing sqlite3 directly. scan_progress.py and
suppression.py now delegate to existing data-layer readers instead of
re-parsing JSON inline.

Deletes the score_run.py:16 and evaluation_mixin.py:366 baseline entries
(services no longer imports analysis/); hand-edits the tooling_mixin.py
line-number shift from the subprocess/urllib import removal; renames the
analysis/_report_scoring.py serialization-boundary key to its new path;
drops the now-stale analysis/cache/local.py logging-ratchet entry.

LocalFileBackend is deliberately NOT re-exported via services/_wiring.py:
its module reaches the top-level quodeq package (CacheEntry.quodeq_version
-> quodeq.__version__), which deferred-imports quodeq.cli -> httpx/pydantic.
_wiring is imported by nearly every services module including
services/scoring (reachable from tests/core), so wiring it there would
make httpx/pydantic transitively reachable from inner-layer tests.
evaluation_mixin.py keeps a narrowly-scoped deferred import instead.
